Relocate `GraphQLConverter` and rename it to `ConversionsToGraphQL`.
